requirements.txt
# YouTube Video Intelligence Pipeline - Requirements
# Install with: pip install -r requirements.txt

# Google APIs
google-api-python-client>=2.100.0
google-generativeai>=0.3.0

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0

# YouTube audio download
yt-dlp>=2024.1.0

# Audio transcription (Whisper)
transformers>=4.36.0
torch>=2.0.0

# Report generation
python-docx>=1.0.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress bars
tqdm>=4.66.0
